backend/requirements.txt
acres==0.5.0
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.14
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.7.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.67
langchain-google-genai==2.1.6
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pathlib==1.0.1
proto-plus==1.26.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyMuPDF==1.26.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==7.1.4
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.2
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zipp==3.23.0
zstandard==0.23.0
